Extract client factory
