+++ sources/Google.Solutions.Apis/Auth/Gaia/GaiaOidcClient.cs
@@ -358,7 +358,7 @@ namespace Google.Solutions.Apis.Auth.Gaia
                       new Uri(directions.BaseUri, "/token").ToString(),
                       new Uri(directions.BaseUri, "/revoke").ToString())
             {
-                this.HttpClientFactory = new Initializers.PscAndMtlsAwareHttpClientFactory(
+                this.HttpClientFactory = new PscAndMtlsAwareHttpClientFactory(
                     directions,
                     deviceEnrollment);
