+++ sources/Google.Solutions.Apis/Client/Initializers.cs
@@ -64,111 +64,5 @@ namespace Google.Solutions.Apis.Client
                     authorization)
             };
         }
-
-        //---------------------------------------------------------------------
-        // Inner classes.
-        //---------------------------------------------------------------------
-
-        /// <summary>
-        /// Client factory that enables client certificate and adds PSC-style Host headers
-        /// if needed.
-        /// </summary>
-        internal class PscAndMtlsAwareHttpClientFactory : IHttpClientFactory, IHttpExecuteInterceptor
-        {
-            private readonly ServiceEndpointDirections directions;
-            private readonly IDeviceEnrollment deviceEnrollment;
-            private readonly ICredential credential;
-
-            public PscAndMtlsAwareHttpClientFactory(
-                ServiceEndpointDirections directions,
-                IAuthorization authorization)
-            {
-                this.directions = directions;
-                this.deviceEnrollment = authorization.DeviceEnrollment;
-                this.credential = authorization.Credential;
-            }
-
-            public PscAndMtlsAwareHttpClientFactory(
-                ServiceEndpointDirections directions,
-                IDeviceEnrollment deviceEnrollment)
-            {
-                this.directions = directions;
-                this.deviceEnrollment = deviceEnrollment;
-                this.credential = null;
-            }
-
-            public ConfigurableHttpClient CreateHttpClient(CreateHttpClientArgs args)
-            {
-                if (this.credential != null)
-                {
-                    args.Initializers.Add(this.credential);
-                }
-
-                var factory = new MtlsAwareHttpClientFactory(this.directions, this.deviceEnrollment);
-                var httpClient = factory.CreateHttpClient(args);
-
-                httpClient.MessageHandler.AddExecuteInterceptor(this);
-
-                return httpClient;
-            }
-
-            public Task InterceptAsync(
-                HttpRequestMessage request,
-                CancellationToken cancellationToken)
-            {
-                if (this.directions.Type == ServiceEndpointType.PrivateServiceConnect)
-                {
-                    Debug.Assert(!string.IsNullOrEmpty(this.directions.Host));
-
-                    //
-                    // We're using PSC, the so hostname we're using to connect is
-                    // different than what the server expects.
-                    //
-                    Debug.Assert(request.RequestUri.Host != this.directions.Host);
-
-                    //
-                    // Inject the normal hostname so that certificate validation works.
-                    //
-                    request.Headers.Host = this.directions.Host;
-                }
-
-                return Task.CompletedTask;
-            }
-        }
-
-        /// <summary>
-        /// Client factory that enables client certificate authenticateion
-        /// if the device is enrolled.
-        /// </summary>
-        internal class MtlsAwareHttpClientFactory : HttpClientFactory
-        {
-            private readonly ServiceEndpointDirections directions;
-            private readonly IDeviceEnrollment deviceEnrollment;
-
-            public MtlsAwareHttpClientFactory(
-                ServiceEndpointDirections directions,
-                IDeviceEnrollment deviceEnrollment)
-            {
-                this.directions = directions.ExpectNotNull(nameof(directions));
-                this.deviceEnrollment = deviceEnrollment.ExpectNotNull(nameof(deviceEnrollment));
-            }
-
-            protected override HttpClientHandler CreateClientHandler()
-            {
-                var handler = base.CreateClientHandler();
-
-                if (this.directions.UseClientCertificate &&
-                    HttpClientHandlerExtensions.CanUseClientCertificates)
-                {
-                    Debug.Assert(this.deviceEnrollment.State == DeviceEnrollmentState.Enrolled);
-                    Debug.Assert(this.deviceEnrollment.Certificate != null);
-
-                    var added = handler.TryAddClientCertificate(this.deviceEnrollment.Certificate);
-                    Debug.Assert(added);
-                }
-
-                return handler;
-            }
-        }
     }
 }
