+++ sources/Google.Solutions.IapDesktop/Properties/launchSettings.json
@@ -1,7 +1,8 @@
 {
   "profiles": {
     "Google.Solutions.IapDesktop": {
-      "commandName": "Project"
+      "commandName": "Project",
+      "commandLineArgs": "/profile Fred"
     }
   }
 }
